Move `year` module to be submodule of `input`
It's not used outside of `input`, so let's keep it in here along with all the other input-related code.
