+++ src/input/agent/search_space.rs
@@ -3,8 +3,8 @@ use super::super::{input_err_msg, read_csv_optional, try_insert};
 use crate::agent::{Agent, AgentID, AgentMap, AgentSearchSpaceMap};
 use crate::commodity::CommodityID;
 use crate::id::IDCollection;
+use crate::input::parse_year_str;
 use crate::process::{Process, ProcessMap};
-use crate::year::parse_year_str;
 use anyhow::{Context, Result};
 use itertools::Itertools;
 use serde::Deserialize;
