+++ src/lib.rs
@@ -26,7 +26,6 @@ pub mod settings;
 pub mod simulation;
 pub mod time_slice;
 pub mod units;
-pub mod year;
 
 #[cfg(test)]
 mod fixture;
